+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.0.1] - 2020-04-20
+
+### Changed
+  - Reimplementation of required, recommended and optional metadata JSON fields
+  - added `acq` to `eeg` modality (present in BIDS 1.2.0)
+
+### Added
+  - `ieeg` and `meg` modalities to `EEG` module
+
 ## [1.0.0] - 2020-04-08
 
 ### Added